✨feature: by id added into user repo

+++ internal/repository/user_repository.go
@@ -8,7 +8,7 @@ import (
 type UserRepository interface {
 	Save(*entity.User) error
 	ByMobileNumber(string, *entity.User) *gorm.DB
-	ById(int64, *entity.User) *gorm.DB
+	ById(uint, *entity.User) *gorm.DB
 }
 
 type userRepository struct {
